Changes and extra Art Assets
I have made the changes suggested by Ethan. I have also created my own iteration of the Sparta Cat, seeing that Toby's was lacking to say the least and used it to express the concept of the background. I have also uploaded a finished version of my first sketch on the theme (that i worked on in my free time, not logging the work on Jira) and found a couple of fonts we could use