Tidy up and Organize
I have organized the meeting minutes and other folders into "Group 14" subfolders that contain part of my work on that team and "Indpendent" folders for my work from now on. I have deleted Toby's artwork part, as I will not be using it and kept only the art that I worked on. I have also Isolated the Group 14 Prototype.